+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Release Candidate
 
+### Features
+
+- Select all feature for credentials in bulk assignment screen. (INDIGO Sprint 230713, [!47](https://github.com/TeskaLabs/seacat-admin-webui/pull/47))
+
 ## v23.29-alpha
 
 ### Breaking change